Organize imports project wide
This follows IntelliJ's formatting style for imports
The next commit will add an editorconfig. The diff for the editor config introduction is not readable because imports are not consistently formatted.

Thus, this commit allows the next commit's diff to be readable.
Some files had CRLF endings and have been converted to LF